feat(stays): rename types file to StaysTypes
