+++ src/types/index.ts
@@ -16,4 +16,4 @@ export * from '../Places/Suggestions/SuggestionsType'
 export * from './shared'
 export * from './ClientType'
 export * from '../notifications/Webhooks/WebhooksType'
-export * from '../Stays/types'
+export * from '../Stays/StaysTypes'
